backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
httpx>=0.26.0
aiofiles>=23.2.0
jinja2>=3.1.0

# Database & Auth
supabase>=2.0.0

# AI/ML
google-generativeai>=0.3.0
elevenlabs>=1.0.0

# Paper Ingestion
arxiv>=2.1.0
pikepdf>=8.0.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
